All backend services specs refactored to individual tests/files
