+++ spec/smart_app_launch/backend_services_authorization_request_success_test_spec.rb
@@ -1,11 +1,7 @@
-require 'pry'
-require_relative '../../lib/smart_app_launch/backend_services_authorization_group'
 require_relative '../../lib/smart_app_launch/backend_services_authorization_request_success_test'
 require_relative '../../lib/smart_app_launch/backend_services_authorization_request_builder'
 
 RSpec.describe SMARTAppLaunch::BackendServicesAuthorizationRequestSuccessTest do
-  include Rack::Test::Methods
-  include RequestHelpers
   
   let(:test) { Inferno::Repositories::Tests.new.find('smart_backend_services_auth_request_success') }
   let(:session_data_repo) { Inferno::Repositories::SessionData.new }
@@ -48,7 +44,6 @@ RSpec.describe SMARTAppLaunch::BackendServicesAuthorizationRequestSuccessTest do
     test_run_params = { test_session_id: test_session.id }.merge(runnable.reference_hash)
     test_run = Inferno::Repositories::TestRuns.new.create(test_run_params)
     inputs.each do |name, value|
-      # binding.pry
       session_data_repo.save(
         test_session_id: test_session.id,
         name: name,
